# API Interface/user-manager-swagger (14-11-2024).yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8093
    description: Generated server url
paths:
  /api/users/update:
    put:
      tags:
        - user-manager-controller
      summary: Update user's information in Keycloak
      operationId: updateUser
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        '200':
          description: User updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
        '400':
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Unable to update user in Keycloak
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/change-password:
    put:
      tags:
        - user-manager-controller
      summary: Change user's password in Keycloak
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        '200':
          description: User's password updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
        '400':
          description: Password is missing
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Unable to update user's password in Keycloak
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/refresh-token:
    post:
      tags:
        - user-manager-controller
      summary: Authenticate user given credentials or refresh token
      operationId: authenticateOrRefreshToken
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsDTO'
      responses:
        '200':
          description: Authentication token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponseDTO'
        '400':
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '401':
          description: Authentication process failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoAuthenticationResponseDTO'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/authenticate:
    post:
      tags:
        - user-manager-controller
      summary: Authenticate user given credentials or refresh token
      operationId: authenticateOrRefreshToken_1
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsDTO'
      responses:
        '200':
          description: Authentication token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponseDTO'
        '400':
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '401':
          description: Authentication process failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoAuthenticationResponseDTO'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/logout:
    post:
      tags:
        - user-manager-controller
      summary: Logout user
      operationId: logoutUser
      responses:
        '200':
          description: User logged out successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
        '400':
          description: 'Invalid request: Either credentials or token must be provided!'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '401':
          description: Authentication process failed!
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/create:
    post:
      tags:
        - user-manager-controller
      summary: Create a new user in Keycloak
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        '200':
          description: User created successfully in Keycloak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponseDTO'
        '400':
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '401':
          description: Authentication process failed!
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '417':
          description: User already exists in Keycloak
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
        '500':
          description: Unable to create user in Keycloak
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users:
    get:
      tags:
        - user-manager-controller
      summary: Retrieve all users from Keycloak
      operationId: fetchUsers
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: An unexpected error occured
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/search:
    get:
      tags:
        - user-manager-controller
      summary: Search user by ID from Keycloak
      operationId: fetchUser
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: An unexpected error occured
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/ids:
    get:
      tags:
        - user-manager-controller
      summary: Retrieve all user IDs from Keycloak
      operationId: getAllUserIds
      responses:
        '200':
          description: User IDs retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '400':
          description: An unexpected error occured
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/ids/role/{realmRole}:
    get:
      tags:
        - user-manager-controller
      summary: Retrieve all user IDs from Keycloak
      operationId: getAllUserIdsByUserRole
      parameters:
        - name: realmRole
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User IDs for role retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '400':
          description: An unexpected error occured
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Unable to locate requested client ID in Keycloak
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/users/auth-info:
    get:
      tags:
        - user-manager-controller
      summary: Retrieve Authentication Information based on the JWT token
      operationId: getAuthInfo
      responses:
        '200':
          description: Information about Authentication Information based on the JWT token
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
...
components:
  schemas:
    ApiResponseInfoMapStringString:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: string
        errors:
          type: object
        message:
          type: string
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
    ApiResponseInfoString:
      type: object
      properties:
        data:
          type: string
        errors:
          type: object
        message:
          type: string
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
    UserDTO:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        userRole:
          type: string
          enum:
            - ROBOT_PROGRAMMER
            - VIRTUAL_COMMISSIONING_ENGINEER
            - ROBOT_SIMULATION_USER
            - PROCESS_ENGINEER
            - PLANT_OPERATOR
            - TECHNICIAN
            - PRODUCTION_SCHEDULING_TEAM_MEMBER
            - MAINTENANCE_ENGINEER
            - INNOVATION_ENGINEER
            - OPERATOR
            - SHOP_FLOOR_MANAGER
            - PLANT_MANAGER
            - PRODUCTION_MANAGER
            - LOGISTICS_MANAGER
            - PRODUCTION_SCHEDULER
            - DESIGN_EXPERT
            - AUTOMATION_MANAGER
            - SUPER_ADMIN
            - NONE
        pilotRole:
          type: string
          enum:
            - USER
            - ADMIN
            - SUPER_ADMIN
            - NONE
        pilotCode:
          type: string
          enum:
            - CRF
            - ILTAR
            - SEW
            - FFT
            - NONE
    CredentialsDTO:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    AuthenticationResponseDTO:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          format: int32
        tokenType:
          type: string
        refreshToken:
          type: string
        refreshExpiresIn:
          type: integer
          format: int32
...